Dirty Fix of animation of Years to the correct Year
Sorry that this is so ugly but what this fixes is: 
The snap function being called multiple times constantly after switching the editor view, this is now handled by animating[method]={lastValue:value} flag which prevents subsequent calls if it will be passing the same value into the function(obviously you have nothing to animate to if it is scrolling to the same value.)

This also has a fix to use practically the same implementation you used to animate the hours and minutes as well as preventing overscroll. 
This is where it gets buggy- There seems to be an issue where no matter what I try I can't seem to center the roledex and when scrolling on the years roledex it seems kind of biased to move downward than upward and sometimes skips values easily. 
If I get it to work I'll send you another pull request